Creates version namespace precedent.
Greatly simplifies the code by removing linked data complexities.
