--- app/controllers/api/services_controller.rb
@@ -1,9 +1,0 @@
-module Api
-  class ServicesController < BaseController
-
-    def contacts
-      render json: { organizations: jsonld_for_collection(Organization.all) }
-    end
-
-  end
-end
