+++ app/controllers/api/v1/organizations_controller.rb
@@ -1,4 +1,4 @@
-module Api
+module Api::V1
   class OrganizationsController < ModelController
 
   end
