--- app/models/api/ability.rb
@@ -1,9 +1,0 @@
-module Api
-  class Ability < ::BaseAbility
-
-    def post_initialize
-      can [:show, :index], :all
-    end
-
-  end
-end
