--- spec/controllers/api/organizations_controller_spec.rb
@@ -1,9 +1,0 @@
-require "rails_helper"
-require "support/shared_examples_for_api_controllers"
-require "support/shared_examples_for_api_model_controllers"
-
-module Api
-  RSpec.describe OrganizationsController, type: :controller do
-    it_behaves_like "an API model controller"
-  end
-end
